backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
anthropic
python-dotenv
psycopg2-binary
